separate statistics value storage classes in perf_def.hpp
