+++ src/thread/thread_daemon.hpp
@@ -28,7 +28,7 @@
 #include "thread_task.hpp"
 #include "thread_waiter.hpp"
 
-#include "perf.hpp"
+#include "perf_def.hpp"
 
 #include <thread>
